Finished implementing adding, editing and removing of both books and authors. starting on inventory.

+++ DB_Labb2/Model/BookAuthor.cs
@@ -8,7 +8,7 @@ namespace DB_Labb2.Model
 {
     public class BookAuthor : ModelBase
     {
-        public int ISBN13 { get; set; }
+        public long ISBN13 { get; set; }
         public int AuthorID { get; set; }
 
         public Book Book { get; set; }
